Add tests for view updating
