+++ VoiceOver Designer/Features/Tests/EditorTests/Base/EditorAfterDidLoadTests.swift
@@ -1,0 +1,9 @@
+@testable import Editor
+
+class EditorAfterDidLoadTests: EditorTests {
+    override func setUp() {
+        super.setUp()
+        
+        didLoad()
+    }
+}
